requirements.txt
requests
psycopg2-binary
python-telegram-bot